+++ src/containerapp/azext_containerapp/_clients.py
@@ -18,6 +18,7 @@ from azure.cli.command_modules.containerapp._clients import (
     GitHubActionClient,
     ContainerAppClient,
     ContainerAppsJobClient,
+    DaprComponentClient,
     ManagedEnvironmentClient,
     WorkloadProfileClient)
 
@@ -492,3 +493,7 @@ class ConnectedEnvStorageClient():
                 env_list.append(formatted)
 
         return env_list
+
+
+class DaprComponentPreviewClient(DaprComponentClient):
+    api_version = PREVIEW_API_VERSION